fix: split composite BootOrder in read patch
